issue #336: file-backed mmap の file offset も long 化 (alloc_and_map / AllocInfo.map_offset)
#336 の残スコープ。mmap の off_t は FileSeek 手前の alloc_and_map(int offset) /
AllocInfo.map_offset(int) で既に切り詰められていた (file-backed mmap を >2GB offset で
張ると file[0..] を誤って map してしまう)。

- MemoryBackend.alloc_and_map の 3 overload の offset を long に。
- Memory / NativeMemoryBackend の alloc_and_map 実装と AllocInfo.map_offset を long に。
- amd64_mmap の (int)offset cast を除去。
- size / map_size は byte[] の 2GB 上限ゆえ据置、srcoff は region 内 offset (<2GB) ゆえ
  据置、i386 sys_mmap は int offset を auto-promote (32-bit ABI、無改変)。

回帰: 新規 sys_mmap_large64 (4GiB offset の file-backed mmap、byte0='M' を確認、sparse)
が software / native(KVM) 両方で PASS。hello_dyn64 (ld.so の libc mmap) native PASS。
これで #336 完了 (lseek/pread64/pwrite64/copy_file_range + file-backed mmap offset)。

+++ src/main/java/emulin/MemoryBackend.java
@@ -96,13 +96,13 @@ public interface MemoryBackend {
   /** anonymous な領域を size byte 確保し、guest 仮想 address を返す。 */
   long    alloc        ( long adrs, int size );
   /** mmap の中核。fd >= 0 で file backed、< 0 で anonymous。prot は次の overload。 */
-  long    alloc_and_map( long adrs, int size, int fd, int offset );
+  long    alloc_and_map( long adrs, int size, int fd, long offset );
   /** prot 指定 ver。PROT_READ / PROT_WRITE / PROT_EXEC を反映 (software は記録のみ)。 */
-  long    alloc_and_map( long adrs, int size, int fd, int offset, int prot );
+  long    alloc_and_map( long adrs, int size, int fd, long offset, int prot );
   /** flags 指定 ver。MAP_FIXED の有無で adrs の意味が変わる (FIXED=その仮想に必ず map、
    *  無し=hint。Linux は hint 範囲が塞がっていると別の場所を選ぶ)。default は flags を
    *  無視して従来挙動 (software backend は既存の byte-identical 挙動を維持)。 */
-  default long alloc_and_map( long adrs, int size, int fd, int offset, int prot, long flags ) {
+  default long alloc_and_map( long adrs, int size, int fd, long offset, int prot, long flags ) {
     return alloc_and_map( adrs, size, fd, offset, prot );
   }
   /** 大きな anonymous 領域を一気に確保 (huge page emulation、glibc malloc 大物用)。 */
